refactor: share tax breakdown helper
